Added a timeout test
