+++ .github/workflows/haskell-ci.yml
@@ -114,6 +114,8 @@ jobs:
           echo "ARG_BENCH=--enable-benchmarks" >> "$GITHUB_ENV"
           echo "HEADHACKAGE=false" >> "$GITHUB_ENV"
           echo "ARG_COMPILER=--$HCKIND --with-compiler=$HC" >> "$GITHUB_ENV"
+          echo "ARG_NORMAL_TEST_SUITES=unit-tests" >> "$GITHUB_ENV"
+          echo "ARG_TIMEOUT_TEST_SUITS=timeout-test-131" >> "$GITHUB_ENV"
         env:
           HCKIND: ${{ matrix.compilerKind }}
           HCNAME: ${{ matrix.compiler }}
@@ -208,15 +210,18 @@ jobs:
         run: |
           $CABAL v2-build $ARG_COMPILER --disable-tests --disable-benchmarks --dependencies-only -j2 all
           $CABAL v2-build $ARG_COMPILER $ARG_TESTS $ARG_BENCH --dependencies-only -j2 all
-      - name: build w/o tests
+      - name: build normal w/o tests
         run: |
-          $CABAL v2-build $ARG_COMPILER --disable-tests --disable-benchmarks all
-      - name: build
+          $CABAL v2-build $ARG_COMPILER --disable-tests --disable-benchmarks $ARG_NORMAL_TEST_SUITES
+      - name: build normal
         run: |
-          $CABAL v2-build $ARG_COMPILER $ARG_TESTS $ARG_BENCH all --write-ghc-environment-files=always
-      - name: tests
+          $CABAL v2-build $ARG_COMPILER $ARG_TESTS $ARG_BENCH $ARG_NORMAL_TEST_SUITES --write-ghc-environment-files=always
+      - name: normal tests
         run: |
-          $CABAL v2-test $ARG_COMPILER $ARG_TESTS $ARG_BENCH all --test-show-details=direct
+          $CABAL v2-test $ARG_COMPILER $ARG_TESTS $ARG_BENCH $ARG_NORMAL_TEST_SUITES --test-show-details=direct
+      - name: compile timeout tests
+        run: |
+          timeout 3m $CABAL v2-build $ARG_COMPILER $ARG_TESTS $ARG_BENCH $ARG_TIMEOUT_TEST_SUITES --write-ghc-environment-files=always
       - name: cabal check
         run: |
           cd ${PKGDIR_ghc_typelits_natnormalise} || false
